Clients info are filed here https://drive.google.com/drive/folders/1IxuDBR22XIlHw96kuaePqfyZPPJIQR8l?usp=drive_link
Cases are filed first by year case was accepted by team, followed by Clients Name (to copy to this system) which are also the google folder names. Inside you will find a google doc file with title including the words "Internal Doc..." - that is the client's session notes file.

It is recorded in a chronological order with a simple tabled header of Session number, attending clinician, date/time, sometimes duration, and optional Location info. We want to use this info to populate our Clients list and Session Notes in this system.

Under it, you will find the clinicians' notes. We want to import that as well and re-record as is as individual sessions, date/time stamp to the original creation date/time in our system too.

Entry versioning history - it will be great if we can sync this too into our system so that the migration of use from current google drive to here is very smooth.